Finished cancel ride page UI
