+++ lib/Home.dart
@@ -8,7 +8,6 @@ import 'package:carriage_rider/Profile.dart';
 import 'package:carriage_rider/Upcoming.dart';
 import 'package:flutter/material.dart';
 import 'package:carriage_rider/Ride_History.dart';
-import 'package:carriage_rider/Upcoming_Ride.dart';
 import 'package:carriage_rider/Current_Ride.dart';
 import 'package:provider/provider.dart';
 import 'package:carriage_rider/Settings.dart';
@@ -109,13 +108,15 @@ class Home extends StatelessWidget {
                   onTap: () {
                     Navigator.push(context,
                         //TODO: remove temporary ride
-                        new MaterialPageRoute(builder: (context) => Upcoming(
+                        new MaterialPageRoute(builder: (context) => UpcomingRidePage(
                             Ride(
                               type: 'active',
                               startLocation: 'Uris Hall',
+                              startAddress: '100 Carriage Way, Ithaca, NY 14850',
                               endLocation: 'Cascadilla Hall',
-                              startTime: DateTime(2020, 10, 18, 12, 0),
-                              endTime: DateTime(2020, 10, 18, 12, 15),
+                              endAddress: '101 DTI St, Ithaca, NY 14850',
+                              startTime: DateTime(2020, 10, 18, 13, 0),
+                              endTime: DateTime(2020, 10, 18, 13, 15),
                             )
                         )));
                   },
