+++ lib/TextThemes.dart
@@ -17,18 +17,18 @@ class TextThemes {
   static final monthStyle = TextStyle(
     color: Colors.black,
     fontWeight: FontWeight.w700,
-    fontSize: 22,
+    fontSize: 22
   );
 
   static final dayStyle = TextStyle(
-    color: Colors.grey[700],
-    fontWeight: FontWeight.w500,
-    fontSize: 22,
+    color: Colors.black,
+    fontWeight: FontWeight.w400,
+    fontSize: 22
   );
 
   static final timeStyle = TextStyle(
     color: Colors.grey[500],
     fontWeight: FontWeight.w400,
-    fontSize: 22,
+    fontSize: 22
   );
 }
